Changes made in the code according to the instructions
