+++ src/pages/CurrentProjectDetails/CurrentProjectDetails.tsx
@@ -1,6 +1,7 @@
 import { ArrowLeft, Building, Calendar, MapPin } from 'lucide-react';
 import * as React from 'react';
 import Header from '../../components/Header/Header';
+import { currentProjectDetailsMessages } from '../../constants/messages';
 import './CurrentProjectDetails.css';
 
 interface Role {
@@ -110,65 +111,62 @@ const CurrentProjectDetails: React.FC = () => {
 
         <div className="project-title-section">
           <div className="title-info">
-            <h1 className="main-project-title">The Dark Knight Returns</h1>
+            <h1 className="main-project-title">{currentProjectDetailsMessages.projectTitle}</h1>
             <div className="project-metadata">
-              <span className="status-badge-active">Active</span>
+              <span className="status-badge-active">{currentProjectDetailsMessages.statusActive}</span>
               <div className="meta1-item">
                 <MapPin size={16} color="#FF7F00" />
-                Los Angeles, California
+                {currentProjectDetailsMessages.location}
               </div>
               <div className="meta1-item">
                 <Calendar size={16} color="#FF7F00" />
-                Deadline: 2021-08-15
+                {currentProjectDetailsMessages.deadline}
               </div>
             </div>
           </div>
           <button className="edit2-button">
-            Edit Project
+            {currentProjectDetailsMessages.editProjectButton}
           </button>
         </div>
         
         <div className="project-details-card">
           <h4 className="section1-title">
             <Building color="#FF7F00" size={20} /> 
-            Project Details
+            {currentProjectDetailsMessages.projectDetailsTitle}
           </h4>
           
           <div className="details-content">
             <div className="details-grid">
               <div className="synopsis-section">
-                <h3 className="subsection-title">Synopsis</h3>
+                <h3 className="subsection-title">{currentProjectDetailsMessages.synopsisTitle}</h3>
                 <p className="synopsis-text">
-                  A gripping tale of justice and redemption in a dark urban setting.
+                  {currentProjectDetailsMessages.synopsisText}
                 </p>
                 <p className="synopsis-description">
-                  In a corrupt, decaying city, a disgraced detective teams up with a wrongfully 
-                  accused woman to uncover the truth behind a brutal murder. As they dig deeper, 
-                  they face a web of lies, betrayal, and moral ambiguity. A gritty tale of 
-                  justice, redemption, and the cost of doing what's right in a world gone dark.
+                  {currentProjectDetailsMessages.synopsisDescription}
                 </p>
               </div>
               
               <div className="production-section">
                 <div className="production-header">
-                  <h3 className="subsection-title">Production Details</h3>
+                  <h3 className="subsection-title">{currentProjectDetailsMessages.projectDetailsTitle}</h3>
                   <div className="genre-tags">
-                    <span className="genre-tag action">Action</span>
-                    <span className="genre-tag drama">Drama</span>
+                    <span className="genre-tag action">{currentProjectDetailsMessages.genreValue}</span>
+                    <span className="genre-tag drama">{currentProjectDetailsMessages.genreValue}</span>
                   </div>
                 </div>
                 <div className="production-info">
                   <div className="production-item">
-                    <span className="production-label">Category:</span>
-                    <span className="production-value">Feature Film</span>
+                    <span className="production-label">{currentProjectDetailsMessages.productionLabel}</span>
+                    <span className="production-value">{currentProjectDetailsMessages.productionValue}</span>
                   </div>
                   <div className="production-item">
-                    <span className="production-label">Genre:</span>
-                    <span className="production-value">Action, Drama</span>
+                    <span className="production-label">{currentProjectDetailsMessages.genreLabel}</span>
+                    <span className="production-value">{currentProjectDetailsMessages.genreValue}</span>
                   </div>
                   <div className="production-item">
-                    <span className="production-label">Production:</span>
-                    <span className="production-value">Warner Bros, DC Films</span>
+                    <span className="production-label">{currentProjectDetailsMessages.productionLabel}</span>
+                    <span className="production-value">{currentProjectDetailsMessages.productionValue}</span>
                   </div>
                 </div>
               </div>
@@ -180,31 +178,31 @@ const CurrentProjectDetails: React.FC = () => {
           <div className="roles-section">
             <h2 className="section1-title">
               <Building color="#FF7F00" size={20} /> 
-              Roles (3)
+              {currentProjectDetailsMessages.rolesTitle}
             </h2>
             
             <div className="roles-list">
               {roles.map((role) => (
                 <div key={role.id} className="role-card">
-                  <div className="role-status-badge">Open</div>
+                  <div className="role-status-badge">{currentProjectDetailsMessages.roleStatusOpen}</div>
                   
                   <h3 className="role1-title">{role.title}</h3>
                   
                   <div className="role-details">
                     <div className="role-detail-item">
-                      <span className="detail-label">Gender: </span>
+                      <span className="detail-label">{currentProjectDetailsMessages.genderLabel}</span>
                       <span className="detail-value">{role.gender}</span>
                     </div>
                     <div className="role-detail-item">
-                      <span className="detail-label">Age: </span>
+                      <span className="detail-label">{currentProjectDetailsMessages.ageLabel} </span>
                       <span className="detail-value">{role.age}</span>
                     </div>
                     <div className="role-detail-item">
-                      <span className="detail-label">Compensation: </span>
+                      <span className="detail-label">{currentProjectDetailsMessages.compensationLabel}</span>
                       <span className="detail-value">{role.compensation}</span>
                     </div>
                     <div className="role-detail-item">
-                      <span className="detail-label">Application: </span>
+                      <span className="detail-label">{currentProjectDetailsMessages.applicationLabel}</span>
                       <span className="detail1-value">{role.applications}</span>
                     </div>
                   </div>
@@ -216,7 +214,7 @@ const CurrentProjectDetails: React.FC = () => {
           <div className="applications-section1">
             <h2 className="section1-title">
               <Building color="#FF7F00" size={20} />
-              Applications (4)
+              {currentProjectDetailsMessages.applicationsTitle}
             </h2>
             
             <div className="applications-list">
@@ -234,12 +232,12 @@ const CurrentProjectDetails: React.FC = () => {
                   
                   <div className="application-details">
                     <div className="application-detail-item">
-                      <span className="detail-label">Experience: </span>
+                      <span className="detail-label">{currentProjectDetailsMessages.experienceLabel}</span>
                       <span className="detail-value">{application.experience}</span>
                     </div>
                       <div className="application-bottom-row">
                         <div className="application-detail-item">
-                          <span className="detail-label">Applied: </span>
+                          <span className="detail-label">{currentProjectDetailsMessages.appliedLabel}</span>
                           <span className="detail-value">{application.appliedDate}</span>
                         </div>
                         <div className="application-location">
@@ -249,7 +247,7 @@ const CurrentProjectDetails: React.FC = () => {
                   </div>
                   
                   <button className="review1-button">
-                    Review Application
+                    {currentProjectDetailsMessages.reviewApplicationButton}
                   </button>
                 </div>
               ))}
