--- src/pages/CurrentProjectDetails/MessageCurrentProjectDetails.tsx
@@ -1,112 +1,0 @@
-export const currentProjectDetailsMessages = {
-  // Navigation
-  backToDashboard: "Back to Dashboard",
-  
-  // Project Header
-  projectTitle: "The Dark Knight Returns",
-  statusActive: "Active",
-  location: "Los Angeles, California",
-  deadline: "Deadline: 2021-08-15",
-  editProjectButton: "Edit Project",
-  
-  // Project Details Section
-  projectDetailsTitle: "Project Details",
-  synopsisTitle: "Synopsis",
-  synopsisText: "A gripping tale of justice and redemption in a dark urban setting.",
-  synopsisDescription: "In a corrupt, decaying city, a disgraced detective teams up with a wrongfully accused woman to uncover the truth behind a brutal murder. As they dig deeper, they face a web of lies, betrayal, and moral ambiguity. A gritty tale of justice, redemption, and the cost of doing what's right in a world gone dark.",
-  
-  // Production Details
-  productionDetailsTitle: "Production Details",
-  genreTags: ["Action", "Drama"],
-  categoryLabel: "Category:",
-  categoryValue: "Feature Film",
-  genreLabel: "Genre:",
-  genreValue: "Action, Drama",
-  productionLabel: "Production:",
-  productionValue: "Warner Bros, DC Films",
-  
-  // Roles Section
-  rolesTitle: "Roles (3)",
-  roleStatusOpen: "Open",
-  
-  // Role Details Labels
-  genderLabel: "Gender: ",
-  ageLabel: "Age: ",
-  compensationLabel: "Compensation: ",
-  applicationLabel: "Application: ",
-  
-  // Applications Section
-  applicationsTitle: "Applications (4)",
-  reviewApplicationButton: "Review Application",
-  
-  // Application Details Labels
-  experienceLabel: "Experience: ",
-  appliedLabel: "Applied: ",
-  
-  // Status Text Mappings
-  statusUnderReview: "Under Review",
-  statusShortlisted: "Shortlisted",
-  statusInterviewScheduled: "Interview Scheduled",
-  
-  // Sample Roles Data
-  rolesData: [
-    {
-      id: "1",
-      title: "Lead Actor",
-      gender: "Male",
-      age: "20-35",
-      compensation: "$50,000",
-      applications: 28,
-      status: "open"
-    },
-    {
-      id: "2", 
-      title: "Supporting Actress",
-      gender: "Female",
-      age: "20-35",
-      compensation: "$30,000",
-      applications: 56,
-      status: "open"
-    },
-    {
-      id: "3",
-      title: "Villain",
-      gender: "Male, Female",
-      age: "30-60",
-      compensation: "$30,000",
-      applications: 20,
-      status: "open"
-    }
-  ],
-  
-  // Sample Applications Data
-  applicationsData: [
-    {
-      id: "1",
-      name: "Sarah Johnson",
-      role: "Lead Actor",
-      experience: "5 years",
-      appliedDate: "2024-07-01",
-      location: "Los Angeles, California",
-      status: "under-review"
-    },
-    {
-      id: "2",
-      name: "Micael Chen", 
-      role: "Supporting Actress",
-      experience: "5 years",
-      appliedDate: "2024-07-01",
-      location: "Los Angeles, California",
-      status: "shortlisted"
-    },
-    {
-      id: "3",
-      name: "Anu Kargaonkar",
-      role: "Lead Actor", 
-      experience: "5 years",
-      appliedDate: "2024-07-01",
-      location: "Los Angeles, California",
-      status: "interview-scheduled"
-    }
-  ]
-};
