--- src/pages/ForgotPassword/MessageForgotPassword.tsx
@@ -1,34 +1,0 @@
-export const forgotPasswordMessages = {
-  // Page Title
-  title: "Forgot Password?",
-  
-  // Input Field
-  inputPlaceholder: "Email address or Username",
-  
-  // Navigation
-  backToLoginText: "Back to",
-  loginLinkText: "Log in",
-  
-  // Buttons
-  sendButton: "Send",
-  orDivider: "OR",
-  
-  // Organization Info
-  organizationName: "FilmyAI",
-  tagLine: "Start your journey today!",
-  
-  // Error Messages
-  emptyInputError: "Please enter your email or username.",
-  invalidEmailError: "Please enter a valid email address.",
-  
-  // Success Messages
-  resetLinkSent: "Reset link has been sent to your email.",
-  
-  // Social Media Icons (classes)
-  socialIcons: [
-    "fab fa-google",
-    "fab fa-facebook-f", 
-    "fab fa-instagram",
-    "fa-brands fa-linkedin"
-  ],
-}
