--- src/pages/HeroPage/MessageHeroPage.tsx
@@ -1,110 +1,0 @@
-export const heroMessages = {
-  title: "Hero Page Title",
-  typedStrings: [
-    "Welcome to our platform",
-    "Discover amazing features", 
-    "Join us today"
-  ],
-  diamondCount: 5
-};
-
-export const aboutMessages = {
-  subtitleFirstWord: "About",
-  subtitleSecondWord: "Us",
-  description: "Your about section description goes here. This should describe your company, mission, or purpose."
-};
-
-export const demoMessages = {
-  subtitleFirstWord: "Request",
-  subtitleSecondWord: "Demo",
-  firstNamePlaceholder: "First Name",
-  lastNamePlaceholder: "Last Name", 
-  emailPlaceholder: "Email Address",
-  mobilePlaceholder: "Mobile Number",
-  submitButtonText: "Request Demo",
-  submitButtonClassName: "btn btn-primary"
-};
-
-export const blogMessages = {
-  subtitleFirstWord: "Latest",
-  subtitleSecondWord: "Blog",
-  posts: [
-    {
-      title: "Blog Post Title 1",
-      image: "/path/to/image1.jpg"
-    },
-    {
-      title: "Blog Post Title 2",
-      image: "/path/to/image2.jpg" 
-    },
-    {
-      title: "Blog Post Title 3",
-      image: "/path/to/image3.jpg"
-    }
-  ]
-};
-
-export const chatMessages = {
-  chatIcon: "fas fa-comments",
-  headerTitle: "Chat Support",
-  botMessage: "Hello! How can I help you today?",
-  userMessage: "Hi there! I need some assistance.",
-  inputPlaceholder: "Type your message..."
-};
-
-export const footerMessages = {
-  shapeSrc: "/path/to/footer-shape.svg",
-  shapeAlt: "Footer Shape",
-  logoSrc: "/path/to/logo.png",
-  logoAlt: "Company Logo",
-  copyright: "© 2024 Your Company Name. All rights reserved.",
-  socialMedia: [
-    {
-      iconClass: "fab fa-facebook",
-      url: "https://facebook.com/yourcompany"
-    },
-    {
-      iconClass: "fab fa-twitter",
-      url: "https://twitter.com/yourcompany"
-    },
-    {
-      iconClass: "fab fa-linkedin", 
-      url: "https://linkedin.com/company/yourcompany"
-    },
-    {
-      iconClass: "fab fa-instagram",
-      url: "https://instagram.com/yourcompany"
-    }
-  ],
-  policyLinks: [
-    {
-      text: "Privacy Policy",
-      url: "/privacy-policy",
-      target: "_blank"
-    },
-    {
-      text: "Terms of Service",
-      url: "/terms-of-service", 
-      target: "_blank"
-    },
-    {
-      text: "Cookie Policy",
-      url: "/cookie-policy",
-      target: "_blank"
-    }
-  ]
-};
-
-export const navigationIds = {
-  hero: "hero",
-  about: "about", 
-  demo: "demo",
-  blog: "blog",
-  chat: "chat"
-};
-
-export const altTextMessages = {
-  blogImage: "Blog Image",
-  footerShape: "Footer Shape",
-  companyLogo: "Company Logo"
-};
