--- src/pages/LandingPage/MessageLandingPage.tsx
@@ -1,164 +1,0 @@
-
-export const landingPageMessages = {
-  // Welcome Section
-  welcomeTitle: "Welcome back,",
-  welcomeHighlight: "Anu!",
-  welcomeDescription: "Ready to land your next big role? We've curated the perfect opportunities based on your profile and experience.",
-  rating: "4.8 Rating",
-  auditionsThisMonth: "12 Auditions This Month",
-  location: "Los Angeles, CA",
-  viewOpportunitiesButton: "View New Opportunities",
-  
-  // KPI Cards
-  activeApplicationsLabel: "Active Applications",
-  activeApplicationsValue: 8,
-  activeApplicationsChange: "+2 this week",
-  
-  pendingResponsesLabel: "Pending Responses", 
-  pendingResponsesValue: 5,
-  pendingResponsesChange: "2 Urgent",
-  
-  profileViewsLabel: "Profile Views",
-  profileViewsValue: 142,
-  profileViewsChange: "+18% this month",
-  
-  successRateLabel: "Success Rate",
-  successRateValue: "73%",
-  successRateChange: "+5% improvement",
-  
-  // Recommendations Section
-  recommendationsTitle: "Recommended for You",
-  recommendationsSubtitle: "Curated based on your profile and preferences",
-  viewAllRecommendationsButton: "View All Recommendations",
-  matchText: "Match",
-  requirementsLabel: "Requirements:",
-  applyNowButton: "Apply Now",
-  
-  // Sample Recommendations Data
-  recommendationsData: [
-    {
-      title: "Drive in Manhattan",
-      role: "Supporting Male Lead", 
-      genre: "Drama/Romance",
-      description: "A heartwarming story about second chances in the big city",
-      location: "New York, NY",
-      duration: "3 months",
-      budget: "$5M - $15M",
-      match: "95%",
-      tags: ["Age 25-35", "Strong drama skills", "NYC local greeted"],
-      image: "/frontend/images/landingPage/lens.png"
-    },
-    {
-      title: "The Last Detective",
-      role: "Detective Partner",
-      genre: "Crime Thriller", 
-      description: "A gritty crime thriller set in modern-day Los Angeles.",
-      location: "Los Angeles, CA",
-      duration: "4 months",
-      budget: "$15M - $20M",
-      match: "82%",
-      tags: ["Age 30-90", "Action experience", "Physical Fitness required"],
-      image: "/frontend/images/landingPage/lens.png"
-    },
-    {
-      title: "Summer's End",
-      role: "Father Figure",
-      genre: "Crime Thriller",
-      description: "An indie film about family bonds and growing up.",
-      location: "Portland, OR", 
-      duration: "6 weeks",
-      budget: "$2M - $6M",
-      match: "82%",
-      tags: ["Age 35-55", "Fatherly personality", "Indie film experience"],
-      image: "/frontend/images/landingPage/lens.png"
-    }
-  ],
-  
-  // Casting Calls Section
-  castingCallsTitle: "Your Casting Calls",
-  castingCallsSubtitle: "Invitations and auditions you've received",
-  viewAllCastingCallsButton: "View All Casting Calls",
-  urgentLabel: "Urgent",
-  directorLabel: "Dir:",
-  dueLabel: "Due:",
-  requiredLabel: "Required:",
-  confirmAuditionButton: "Confirm Audition",
-  viewDetailsButton: "View Details",
-  
-  // Sample Casting Calls Data
-  castingCallsData: [
-    {
-      title: "Ocean's Revenge",
-      role: "Tech Specialist",
-      urgent: true,
-      director: "James Whpn",
-      date: "Dec 15, 2024",
-      time: "2:30PM–3:30PM", 
-      due: "Dec 10, 2024",
-      required: ["1-min Monologue"],
-      description: "Seeking a charismatic actor for a tech-savvy character in action thriller."
-    },
-    {
-      title: "Shadows & Lights",
-      role: "Lead Detective",
-      urgent: false,
-      director: "Ava Moren",
-      date: "Jan 5, 2025",
-      time: "1:00PM–2:00PM",
-      due: "Dec 30, 2024", 
-      required: ["Headshot", "Resume"],
-      description: "Seeking a charismatic actor for a tech-savvy character in action thriller."
-    }
-  ],
-  
-  // Categories Section
-  categoriesTitle: "Browse by Categories",
-  categoriesSubtitle: "Explore opportunities across different types of productions and genres",
-  availableRoles: "available roles",
-  browseRolesButton: "Browse Roles",
-  viewAllCategoriesButton: "View All Categories",
-  
-  // Categories Data
-  categoriesData: [
-    {
-      label: "Feature Films",
-      roles: 8,
-      description: "Major studio and independent feature films"
-    },
-    {
-      label: "TV Series", 
-      roles: 24,
-      description: "Television series and streaming shows"
-    },
-    {
-      label: "Commercials",
-      roles: 45,
-      description: "Brand commercials and advertising campaigns"
-    },
-    {
-      label: "Web Series",
-      roles: 10,
-      description: "Online content and digital series"
-    },
-    {
-      label: "Romance",
-      roles: 5,
-      description: "Romantic comedies and drama films"
-    },
-    {
-      label: "Action/Thriller",
-      roles: 50,
-      description: "Action-packed and suspenseful productions"
-    },
-    {
-      label: "Comedy",
-      roles: 24,
-      description: "Comedy films and lighthearted content"
-    },
-    {
-      label: "Horror/Sci-Fi",
-      roles: 12,
-      description: "Horror and science fiction productions"
-    }
-  ]
-};
