--- src/pages/Login/MessageLogin.tsx
@@ -1,32 +1,0 @@
-
-export const loginMessages = {
-  // Page Title
-  title: "Login",
-  
-  // Form Fields
-  emailPlaceholder: "Email address or Username",
-  passwordPlaceholder: "Password",
-  
-  // Links and Navigation
-  forgotPasswordText: "Forgot Password?",
-  loginButtonText: "Log in",
-  signupLinkText: "Don't have an account?",
-  signupText: "Sign Up",
-  orDivider: "OR",
-  
-  // Organization Branding
-  organizationName: "FilmyAI",
-  tagLine: "Start your journey today!",
-  
-  // Error Messages
-  emailRequiredError: "Email is required.",
-  invalidEmailError: "Please enter a valid email address.",
-  passwordRequiredError: "Password is required.",
-  
-  // Routes
-  forgotPasswordRoute: "/Forgot",
-  signupRoute: "/Signup",
-  profileRoute: "/profile",
-  
- 
-};
