--- src/pages/ProductionDashboard/MessageProductionDashboard.tsx
@@ -1,98 +1,0 @@
-export const productionDashboardMessages = {
-  // Dashboard Header
-  dashboardTitle: "Production Dashboard",
-  dashboardSubtitle: "Manage your projects and casting calls",
-  
-  // Current Projects Section
-  currentProjectsTitle: "Current Projects",
-  applicationsLabel: "Applications:",
-  rolesLabel: "Roles:",
-  deadlineLabel: "Deadline:",
-  editButton: "Edit",
-  viewDetailsButton: "View Details",
-  
-  // Project Status
-  statusOpen: "Open",
-  statusClosed: "Closed",
-  
-  // Recent Applications Section
-  recentApplicationsTitle: "Recent applications",
-  
-  // Table Headers
-  applicationTableHeaders: {
-    applications: "Applications",
-    role: "Role", 
-    project: "Project",
-    applicationDate: "Application Date",
-    status: "Status",
-    action: "Action"
-  },
-  
-  // Action Buttons
-  reviewButton: "Review",
-  
-  // Application Status Options
-  statusUnderReview: "Under Review",
-  statusShortlisted: "Shortlisted", 
-  statusInterviewScheduled: "Interview Scheduled",
-  
-  // Sample Current Projects Data
-  currentProjectsData: [
-    {
-      title: "The Dark Knight returns",
-      tags: ["Action", "Drama"],
-      applications: 125,
-      roles: 125,
-      deadline: "08-15-2024",
-      status: "Open"
-    },
-    {
-      title: "Summer Romance",
-      tags: ["Romance", "Comedy"],
-      applications: 98,
-      roles: 50,
-      deadline: "09-01-2024",
-      status: "Open"
-    },
-    {
-      title: "Sci-Fi Adventure",
-      tags: ["Sci-Fi", "Adventure"],
-      applications: 75,
-      roles: 30,
-      deadline: "10-10-2024",
-      status: "Closed"
-    }
-  ],
-  
-  // Sample Recent Applications Data
-  recentApplicationsData: [
-    {
-      name: "Sarah Johnson",
-      role: "Lead Actress",
-      project: "The Dark Knight Returns",
-      date: "08-15-2024",
-      status: "Under Review"
-    },
-    {
-      name: "Michael Chen",
-      role: "Supporting Actor",
-      project: "Summer Romance",
-      date: "06-23-2024",
-      status: "Shortlisted"
-    },
-    {
-      name: "Emma Davis",
-      role: "Villain",
-      project: "The Dark Knight Returns",
-      date: "04-11-2024",
-      status: "Interview Scheduled"
-    },
-    {
-      name: "Anu Kargaonkar",
-      role: "Lead Actress",
-      project: "Summer Romance",
-      date: "01-08-2024",
-      status: "Shortlisted"
-    }
-  ]
-};
