+++ src/pages/ProductionDashboard/ProductionDashboard.tsx
@@ -1,6 +1,7 @@
 import { Building } from 'lucide-react';
 import React from 'react';
 import Header from '../../components/Header/Header';
+import { productionDashboardMessages } from '../../constants/messages';
 import './ProductionDashboard.css';
 
 const ProductionDashboard: React.FC = () => {
@@ -81,12 +82,12 @@ const ProductionDashboard: React.FC = () => {
       <Header />
       <main className="dashboard-main">
         <section className="dashboard-header">
-          <h2>Production Dashboard</h2>
-          <p>Manage your projects and casting calls</p>
+          <h2>{productionDashboardMessages.dashboardTitle}</h2>
+          <p>{productionDashboardMessages.dashboardSubtitle}</p>
         </section>
 
         <section className="projects-section">
-          <h3><Building color="#ff7f00" size={20} /> Current Projects</h3>
+          <h3><Building color="#ff7f00" size={20} /> {productionDashboardMessages.currentProjectsTitle}</h3>
           <div className="projects-grid">
             {currentProjects.map((proj, idx) => (
               <div key={idx} className="project-card">
@@ -101,21 +102,21 @@ const ProductionDashboard: React.FC = () => {
                 </div>
                 <div className="card-info">
                   <div className="info-row">
-                    <span className="info-label">Applications:</span>
+                    <span className="info-label">{productionDashboardMessages.applicationsLabel}</span>
                     <span className="info-value">{proj.applications}</span>
                   </div>
                   <div className="info-row">
-                    <span className="info-label">Roles:</span>
+                    <span className="info-label">{productionDashboardMessages.rolesLabel}</span>
                     <span className="info-value">{proj.roles}</span>
                   </div>
                   <div className="info-row">
-                    <span className="info-label">Deadline:</span>
+                    <span className="info-label">{productionDashboardMessages.deadlineLabel}</span>
                     <span className="info-value">{proj.deadline}</span>
                   </div>
                 </div>
                 <div className="card-actions">
-                  <button className="edit1-btn">Edit</button>
-                  <button className="view-btn">View Details</button>
+                  <button className="edit1-btn">{productionDashboardMessages.editButton}</button>
+                  <button className="view-btn">{productionDashboardMessages.viewDetailsButton}</button>
                 </div>
               </div>
             ))}
@@ -123,7 +124,7 @@ const ProductionDashboard: React.FC = () => {
         </section>
 
         <section className="applications-section">
-          <h3><Building color="#ff7f00" size={20} /> Recent applications</h3>
+          <h3><Building color="#ff7f00" size={20} /> {productionDashboardMessages.recentApplicationsTitle}</h3>
           <table className="applications-table">
             <thead>
               <tr>
@@ -143,7 +144,7 @@ const ProductionDashboard: React.FC = () => {
                   <td>{app.project}</td>
                   <td>{app.date}</td>
                   <td>{renderStatusTag(app.status)}</td>
-                  <td><button className="review-btn">Review</button></td>
+                  <td><button className="review-btn">{productionDashboardMessages.reviewButton}</button></td>
                 </tr>
               ))}
             </tbody>
