--- src/pages/RoleDetailPage/MessageRoleDetailPage.tsx
@@ -1,60 +1,0 @@
-
-export const MessageRoleDetailPage = {
-  // Navigation
-  backToMovieDetails: "Back to Movie Details",
-  
-  // Project Header
-  projectTitle: "Midnight in Manhattan - Drama/Romance",
-  roleTitle: "Supporting Male Lead",
-  characterName: "David Chen",
-  
-  // Meta Information
-  director: "Director: Sarah Chen",
-  applyBy: "Apply by: January 15, 2025",
-  
-  // Main Content Sections
-  characterDescriptionTitle: "Character Description",
-  characterDescription: "A charming and witty architect who becomes Sarah's love interest. David is confident yet vulnerable, with a sharp sense of humor that masks his own past disappointments. He's successful in his career but has been unlucky in love until he meets Sarah. The character requires someone who can balance comedy with genuine emotional depth.",
-  
-  requirementsTitle: "Requirements",
-  requirements: [
-    "Strong dramatic and comedic acting skills",
-    "NYC local preferred",
-    "Previous romantic lead experience",
-    "Comfortable with intimate scenes"
-  ],
-  
-  specialSkillsTitle: "Special Skills Required",
-  specialSkills: [
-    "Piano playing",
-    "Ballroom dancing",
-    "Rock climbing"
-  ],
-  
-  // Apply Section
-  applyButtonText: "Apply for Role",
-  
-  // Sidebar - Role Information
-  roleInformationTitle: "Role Information",
-  
-  ageRangeLabel: "Age Range",
-  ageRangeValue: "25-35 years",
-  
-  genderLabel: "Gender",
-  genderValue: "Male",
-  
-  languagesLabel: "Languages Required",
-  languages: [
-    "English",
-    "Mandarin",
-    "Mandarin", // Note: There are duplicate entries in the original
-    "Mandarin",
-    "Mandarin"
-  ],
-  
-  // Console log messages (for development)
-  consoleMessages: {
-    applyForRole: "Apply for role clicked",
-    backToMovieDetails: "Back to movie details clicked"
-  }
-};
