--- src/pages/SignUp/MessageSignUp.tsx
@@ -1,74 +1,0 @@
-export const signUpMessages = {
-  // Page Title
-  title: "Sign Up",
-  
-  // Form Type Options
-  userOption: "User",
-  businessOption: "Business",
-  
-  // Form Field Placeholders
-  firstNamePlaceholder: "First Name",
-  lastNamePlaceholder: "Last Name",
-  businessNamePlaceholder: "Business Name",
-  emailPlaceholder: "Email address",
-  passwordPlaceholder: "Password",
-  
-  // Terms & Conditions
-  termsText: "I agree to",
-  termsLinkText: "Terms & Conditions",
-  termsUrl: "/documents/Terms.pdf",
-  
-  // Buttons
-  signUpButton: "Sign Up",
-  orDivider: "OR",
-  
-  // Login Prompt
-  loginPromptText: "Already have an account?",
-  loginLinkText: "Click here",
-  loginPromptSuffix: "to login",
-  
-  // Organization Branding
-  organizationName: "FilmyAI",
-  tagLine: "Start your journey today!",
-  
-  // Error Messages
-  nameRequiredError: "First and last name are required.",
-  businessNameRequiredError: "Business name is required.",
-  emailRequiredError: "A valid email address is required.",
-  passwordLengthError: "Password must be at least 6 characters long.",
-  termsRequiredError: "You must accept the Terms & Conditions.",
-  
-  // Success Messages
-  thankYouMessage: "Thank you for signing up!",
-  closeButton: "Close",
-  
-  // Social Media Icons
-  socialIcons: [
-    "fab fa-google",
-    "fab fa-facebook-f",
-    "fab fa-instagram",
-    "fa-brands fa-linkedin"
-  ],
-  
-  // Routes
-  loginRoute: "/Login",
-  landingPageRoute: "/LandingPage",
-  
-  // Form Field Names
-  formFieldNames: {
-    firstName: "firstName",
-    lastName: "lastName", 
-    businessName: "businessName",
-    email: "email",
-    password: "password",
-    termsAccepted: "termsAccepted",
-    type: "type"
-  },
-  
-  // Form Type Values
-  formTypes: {
-    user: "user",
-    business: "business"
-  },
-
-}
